requirements.txt
fastapi
pydantic
uvicorn
debugpy
numpy
tensorflow
sklearn.model_selection
sklearn.preprocessing